+++ src/core/core/service_engine.cpp
@@ -26,7 +26,6 @@
 
 #include "service_engine.h"
 #include "task_engine.h"
-#include "core/aio/disk_engine.h"
 #include "rpc_engine.h"
 
 #include <dsn/utility/filesystem.h>
